rename logrus_syslog package to syslog
Package names shouldn't be using underscores. There is also nothing
wrong with naming this package `syslog`. People can include it with any
name they wish.
