+++ hooks/syslog/README.md
@@ -6,12 +6,12 @@
 import (
   "log/syslog"
   "github.com/sirupsen/logrus"
-  logrus_syslog "github.com/sirupsen/logrus/hooks/syslog"
+  lSyslog "github.com/sirupsen/logrus/hooks/syslog"
 )
 
 func main() {
   log       := logrus.New()
-  hook, err := logrus_syslog.NewSyslogHook("udp", "localhost:514", syslog.LOG_INFO, "")
+  hook, err := lSyslog.NewSyslogHook("udp", "localhost:514", syslog.LOG_INFO, "")
 
   if err == nil {
     log.Hooks.Add(hook)
@@ -25,12 +25,12 @@ If you want to connect to local syslog (Ex. "/dev/log" or "/var/run/syslog" or "
 import (
   "log/syslog"
   "github.com/sirupsen/logrus"
-  logrus_syslog "github.com/sirupsen/logrus/hooks/syslog"
+  lSyslog "github.com/sirupsen/logrus/hooks/syslog"
 )
 
 func main() {
   log       := logrus.New()
-  hook, err := logrus_syslog.NewSyslogHook("", "", syslog.LOG_INFO, "")
+  hook, err := lSyslog.NewSyslogHook("", "", syslog.LOG_INFO, "")
 
   if err == nil {
     log.Hooks.Add(hook)
